requirements.txt
transformers==4.43.3
accelerate==0.33.0
datasets==2.20.0
streamlit==1.34.0
pillow==10.4.0

jinja2==3.1.4
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
boto3==1.34.149
uvicorn==0.30.4
fastapi==0.111.1
seaborn==0.13.2
pandas==2.2.2
matplotlib==3.5.2
sqlalchemy
pymysql
pydantic
httpx
python-jose
bcrypt
numpy==1.26.4
protobuf<5,>=3.20
sentence-transformers==2.2.2
transformers==4.33.2
faiss-cpu==1.11.0
langchain==0.2.x
langchain-huggingface==0.0.x
pymongo
